Add docs and remove duplicate code
