+++ simple_saver/getter.go
@@ -4,9 +4,11 @@ import (
 	"github.com/brunograssano/Distribuidos-TP1/common/communication"
 	dataStructures "github.com/brunograssano/Distribuidos-TP1/common/data_structures"
 	"github.com/brunograssano/Distribuidos-TP1/common/filemanager"
+	"github.com/brunograssano/Distribuidos-TP1/common/utils"
 	log "github.com/sirupsen/logrus"
 )
 
+// Getter Server that waits for clients asking for the pipeline results
 type Getter struct {
 	c       *SaverConfig
 	server  *communication.PassiveTCPSocket
@@ -14,6 +16,7 @@ type Getter struct {
 	canSend chan bool
 }
 
+// NewGetter Creates a new results getter server
 func NewGetter(c *SaverConfig, canSend chan bool) (*Getter, error) {
 	server, err := communication.NewPassiveTCPSocket(c.GetterAddress)
 	if err != nil {
@@ -23,8 +26,9 @@ func NewGetter(c *SaverConfig, canSend chan bool) (*Getter, error) {
 	return &Getter{c: c, server: server, stop: make(chan bool), canSend: canSend}, nil
 }
 
+// ReturnResults Basic server loop to return the results
 func (g *Getter) ReturnResults() {
-	defer dataStructures.CloseSocketAndNotifyError(g.server)
+	defer utils.CloseSocketAndNotifyError(g.server)
 	for {
 		socket, err := g.server.Accept()
 		if err != nil {
@@ -38,22 +42,23 @@ func (g *Getter) ReturnResults() {
 		default:
 			g.askLaterForResults(clientSerializer)
 		}
-
 	}
 }
 
+// askLaterForResults Tells the client to wait and finishes the connection
 func (g *Getter) askLaterForResults(resultsSerializer *dataStructures.ResultsSerializer) {
 	resultsSerializer.AskLaterForResults()
 	resultsSerializer.Close()
 }
 
+// sendResults Sends the saved results to the client
 func (g *Getter) sendResults(resultsSerializer *dataStructures.ResultsSerializer) {
 	defer resultsSerializer.Close()
 	reader, err := filemanager.NewFileReader(g.c.OutputFileName)
 	if err != nil {
 		return
 	}
-	defer closeFile(reader.FileManager)
+	defer utils.CloseFileAndNotifyError(reader.FileManager)
 	for reader.CanRead() {
 		select {
 		case <-g.stop:
@@ -76,8 +81,9 @@ func (g *Getter) sendResults(resultsSerializer *dataStructures.ResultsSerializer
 
 }
 
+// Close Stops the execution of the getter server
 func (g *Getter) Close() {
 	g.stop <- true
 	close(g.stop)
-	dataStructures.CloseSocketAndNotifyError(g.server)
+	utils.CloseSocketAndNotifyError(g.server)
 }
